[refactor/#32] Color 로직 Screen 코드로 이동
